Allow servers to restart on source change and continue working
Hakken's behavior demands that a coordinator be present before your
client starts.

If a hakken client starts before any coordinator is present, it will
delete the coordinator from memory and never use it again.  This means
that if your daemon uses hakken for client-side load balancing, and
your coordinator is starting at the same time, that the service will
never use the coordinator.  This means that even if all services,
including the coordinator are running at the same time, they will
refuse to talk to one another.

Client side load balancing is an anti-pattern, full of problems.  In
order to work around this one, we must ensure that the hakken
coordinator starts before any other daemon.  Introducing a sleep
before restarting or starting those daemons somewhat ameliorates the
problem by giving the Hakken coordinator a chance to initialize itself
before the services assume it does not exist.

The -d 10 switch to nodemon tells nodemon to delay 10 seconds before
restarting the daemon after a source file change has been detected.

A proper solution would be to configure additional styx endpoints as
load balancers for every service that needs load balancing.  Clients
that depend on these services can then safely use the canonical
load-balanced URIs.

It's important to note that the sleep/delay configuration does not
solve the problems mentioned above because hakken's delay behavior is
unpredictable... it typically requires 2 or 3 full "tick" rounds
before the system is stable enough for actual use, even when all
services are correctly configured and running.  This makes for very
unpredictable and hard to debug system behavior.